разделяет написанный ранее код на модули
