+++ js/consts.js
@@ -1,0 +1,1 @@
+export const MAX_COUNT_THUMBNAIL = 25;
